tstv/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
plotly
scipy
python-multipart
requests
openpyxl
motor
python-dotenv
pydantic
httpx
matplotlib